backend/requirements.txt
fastapi[standard]
uvicorn[standard]
ultralytics
pydantic
python-multipart
opencv-python
pillow
numpy
torch
